+++ packages/catalog-react/src/index.ts
@@ -2,3 +2,5 @@ export * from './catalog';
 export * from './renderers/simple-table';
 export * from './renderers/flex-table';
 export * from './renderers/latex-formula';
+export * from './renderers/choice-card';
+export * from './renderers/form-card';