+++ packages/catalog-schemas/src/index.ts
@@ -3,4 +3,6 @@ export * from './types';
 export * from './components/simple-table';
 export * from './components/flex-table';
 export * from './components/latex-formula';
+export * from './components/choice-card';
+export * from './components/form-card';
 export * from './catalog';